requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.42
Pillow==7.2.0
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
termcolor==1.1.0
tweepy==3.9.0
urllib3==1.25.10
